Bug: Customers greeted you twice at the start of levels.
In Career mode, when the puzzle window first appears the customer said hello.
But then when the player presses start, the customer said hello again.

I've tweaked the hello timer so it only runs when a puzzle is active,
and so that it defaults to the minimum value. This ensures the customers
say 'hello' at the start of a puzzle, and then only say hello again if
the player plays the puzzle for a little while.